renamed to connectionTest
