+++ src/test/java/LiveChatAPI/LiveChatTestSuite.java
@@ -11,7 +11,7 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 
 @Suite.SuiteClasses({
-        LiveChatConnectionTest.class,
+        ConnectionTest.class,
         GetIntitalDataTest.class,
         RegisterTest.class,
         LoginTest.class,
